requirements.txt
pillow
pandas
scikit-learn
